Reshuffled the classes a little into new phe.utils.reader module.
ParallelVCFReader can now be found and used to travers multiple VCF
readers one position at a time.
